Add files via upload
This dataset contains data on the stablecoins DAI, WLUNA, PAX, USDC, USDT and USTC, including price info and events data for each stablecoin (name of the event + effect). 

After testing some scatterplot 3d visualisations, WLUNA and USTC look the most interesting because they have some huge drops after some negative events occur, but we could possibly use the other ones to show how the effects of an event on the prices can vary. Also, we don't have to do scatter plots, that was just for me to get the gist of the data, so any other better ones would be good.

Something interesting to do with the VR app is to allow scaling changes since with some tokens the prices are very close together while others range by a lot, especially WLUNA, especially if we want to show them all together simultaneously.